+++ Packages/BrandonUtils/Standalone/Strings/LineStyle.cs
@@ -1,0 +1,7 @@
+﻿namespace BrandonUtils.Standalone.Strings {
+    public enum LineStyle {
+        Dynamic = default,
+        Multi,
+        Single,
+    }
+}