+++ Packages/BrandonUtils/Standalone/Strings/PrettificationFlags.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace BrandonUtils.Standalone.Strings {
-    [Flags]
-    public enum PrettificationFlags {
-        None = 0,
-        IncludeTypeLabels
-    }
-}