+++ Packages/BrandonUtils/Standalone/Strings/Prettifiers/InnerPretty.InnerPretty.IEnumerable.cs
@@ -11,21 +11,20 @@ using JetBrains.Annotations;
 namespace BrandonUtils.Standalone.Strings.Prettifiers {
     internal static partial class InnerPretty {
         public static string PrettifyEnumerableT<T>([NotNull] [ItemCanBeNull] IEnumerable<T> enumerable, [CanBeNull] PrettificationSettings settings) {
-            var lineStyle = settings?.PreferredLineStyle ?? default;
+            var lineStyle = settings?.PreferredLineStyle.Value ?? default;
             return lineStyle switch {
-                PrettificationSettings.LineStyle.Multi   => PrettifyEnumerable_MultiLine(enumerable, enumerable.GetType(), settings),
-                PrettificationSettings.LineStyle.Single  => PrettifyEnumerable_SingleLine(enumerable, enumerable.GetType(), settings),
-                PrettificationSettings.LineStyle.Dynamic => PrettifyEnumerable_DynamicLine(enumerable, enumerable.GetType(), settings),
-                _                                        => throw BEnum.InvalidEnumArgumentException(nameof(PrettificationSettings.PreferredLineStyle), lineStyle)
+                LineStyle.Multi   => PrettifyEnumerable_MultiLine(enumerable, enumerable.GetType(), settings),
+                LineStyle.Single  => PrettifyEnumerable_SingleLine(enumerable, enumerable.GetType(), settings),
+                LineStyle.Dynamic => PrettifyEnumerable_DynamicLine(enumerable, enumerable.GetType(), settings),
+                _                 => throw BEnum.InvalidEnumArgumentException(nameof(PrettificationSettings.PreferredLineStyle), lineStyle)
             };
         }
 
         private static string PrettifyEnumerable_DynamicLine<T>(
             [NotNull] [ItemCanBeNull]
             IEnumerable<T> enumerable,
-            [NotNull] Type enumerableType,
-            [CanBeNull]
-            PrettificationSettings settings
+            [NotNull]   Type                   enumerableType,
+            [CanBeNull] PrettificationSettings settings
         ) {
             enumerable = enumerable.ToList();
             var single = PrettifyEnumerable_SingleLine(enumerable, enumerableType, settings);
@@ -35,53 +34,39 @@ namespace BrandonUtils.Standalone.Strings.Prettifiers {
         private static string PrettifyEnumerable_MultiLine<T>(
             [NotNull] [ItemCanBeNull]
             IEnumerable<T> enumerable,
-            [NotNull] Type enumerableType,
-            [CanBeNull]
-            PrettificationSettings prettificationSettings
+            [NotNull]   Type                   enumerableType,
+            [CanBeNull] PrettificationSettings prettificationSettings
         ) {
             enumerable = enumerable.ToList();
-            var str = enumerable.Select(it => it.Prettify().Indent())
-                                .Prepend("[")
-                                .Append("]")
-                                .JoinLines();
 
-            if (prettificationSettings?.Flags.HasFlag(PrettificationFlags.IncludeTypeLabels) == true) {
-                var typeLabel = enumerableType.PrettifyType();
-                str = $"{typeLabel}{str}";
-            }
-
-            return str;
+            return enumerable.Select(it => it.Prettify().Indent())
+                             .Prepend("[")
+                             .Append("]")
+                             .JoinLines()
+                             .WithTypeLabel(enumerableType, prettificationSettings);
         }
 
         public static string PrettifyEnumerable([NotNull] [ItemCanBeNull] IEnumerable enumerable, [CanBeNull] PrettificationSettings settings) {
             var asObjects      = enumerable.Cast<object>();
             var enumerableType = enumerable.GetType();
-            var lineStyle      = settings?.PreferredLineStyle ?? default;
+            var lineStyle      = settings?.PreferredLineStyle.Value ?? default;
             return lineStyle switch {
-                PrettificationSettings.LineStyle.Dynamic => PrettifyEnumerable_DynamicLine(asObjects, enumerableType, settings),
-                PrettificationSettings.LineStyle.Multi   => PrettifyEnumerable_MultiLine(asObjects, enumerableType, settings),
-                PrettificationSettings.LineStyle.Single  => PrettifyEnumerable_SingleLine(asObjects, enumerableType, settings),
-                _                                        => throw BEnum.InvalidEnumArgumentException(nameof(PrettificationSettings.PreferredLineStyle), lineStyle)
+                LineStyle.Dynamic => PrettifyEnumerable_DynamicLine(asObjects, enumerableType, settings),
+                LineStyle.Multi   => PrettifyEnumerable_MultiLine(asObjects, enumerableType, settings),
+                LineStyle.Single  => PrettifyEnumerable_SingleLine(asObjects, enumerableType, settings),
+                _                 => throw BEnum.InvalidEnumArgumentException(nameof(PrettificationSettings.PreferredLineStyle), lineStyle)
             };
         }
 
         private static string PrettifyEnumerable_SingleLine<T>(
             [NotNull] [ItemCanBeNull]
             IEnumerable<T> enumerable,
-            [NotNull] Type enumerableType,
-            [CanBeNull]
-            PrettificationSettings prettificationSettings
+            [NotNull]   Type                   enumerableType,
+            [CanBeNull] PrettificationSettings prettificationSettings
         ) {
             enumerable = enumerable.ToList();
             var joined = enumerable.Select(it => it.Prettify()).JoinString(", ");
-            var str    = $"[{joined}]";
-
-            if (prettificationSettings?.Flags.HasFlag(PrettificationFlags.IncludeTypeLabels) == true) {
-                var typeLabel = enumerableType.PrettifyType();
-                str = $"{typeLabel}{str}";
-            }
-
-            return str;
+            return $"[{joined}]".WithTypeLabel(enumerableType, prettificationSettings);
         }
     }
 }