+++ Packages/BrandonUtils/Testing/IMultipleAsserter.cs
@@ -1,0 +1,7 @@
+﻿namespace BrandonUtils.Testing {
+    public interface IMultipleAsserter {
+        void Invoke();
+
+        int Indent { get; set; }
+    }
+}